Add file deletion functionality

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -143,6 +144,7 @@ exports.postEditProduct = (req, res, next) => {
 
             product.title = updatedTitle;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             product.price = updatedPrice;
@@ -180,8 +182,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
+            console.log('Product deleted!')
             res.redirect('/admin/product-list');
         })
         .catch(err => {
